pyproject.toml
[tool.poetry]
name = "screenie"
version = "0.1.0"
description = ""
authors = ["Kyle Kelley <rgbkrk@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PyAutoGUI = "^0.9.54"
Pillow = "^10.1.0"
openai = "^1.3.3"
elevenlabs = "^0.2.26"
python-ulid = "^2.2.0"


[tool.poetry.group.dev.dependencies]
types-pillow = "^10.1.0.2"
types-pyautogui = "^0.9.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
